requirements.txt
streamlit>=1.30.0
google-genai>=1.0.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
requests>=2.31.0
openpyxl>=3.1.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
supabase>=2.0.0
